회원 관리 페이지 완료

+++ src/main/java/com/simplane/domain/AuthVO.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class AuthVO {
 
-    private Integer memberid;
+    private Long memberid;
     private String auth;
 }
